pyproject.toml
[tool.poetry]
name = "calculator"
version = "0.1.0"
description = ""
authors = ["Robin Maasjosthusmann <robin@maasjosthusmann.de>"]
license = "BSD3"

[tool.poetry.dependencies]
python = "3.9.7"
jupyterlab = "^3.2.8"
numpy = "^1.22.1"
pandas = "^1.3.5"
pytest = "^6.2.5"
pytest-mock = "^3.6.1"
pytest-parametrization = "^2021.4.1"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
flake8 = "^4.0.1"
flake8-docstrings = "^1.6.0"
mypy = "^0.931"
nox = "^2022.1.7"
pylint = "^2.12.2"
safety = "^1.10.3"
pytest = "^6.2.5"
pytest-timeout = "^2.0.2"
pytest-mock = "^3.6.1"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
